Moved unified components to seperate folder
